requirements.txt
pandas
numpy
scipy
scikit-learn
shap
matplotlib
seaborn
streamlit
plotly
xgboost
sentence-transformers
torch
fpdf2
